+++ src/components/formItem/formItem.css
@@ -1,0 +1,10 @@
+.form-fieldset {
+  border: none;
+}
+
+.form-item {
+  display: flex;
+  flex-direction: column;
+
+  margin-bottom: 36px;
+}
