+++ src/components/tag/caption/Caption.ts
@@ -1,4 +1,5 @@
 import CaptionProps from './CaptionProps';
+import './caption.css';
 
 class Caption extends HTMLSpanElement {
   constructor(props: CaptionProps) {
