+++ src/components/tag/image/image.css
@@ -1,0 +1,4 @@
+.category-icon {
+  width: 36px;
+  height: 36px;
+}
