+++ src/components/tag/select/select.css
@@ -1,0 +1,26 @@
+@import url('../../../styles/style.css');
+
+.form-item select {
+  height: 44px;
+  padding: 8px;
+  margin: 6px 0;
+
+  border: 1px solid var(--grey-200);
+  border-radius: 8px;
+
+  font-size: 16px;
+  color: var(--grey-300);
+}
+
+.restaurant-filter-container select {
+  height: 44px;
+  min-width: 125px;
+
+  border: 1px solid #d0d5dd;
+  border-radius: 8px;
+  background: transparent;
+
+  font-size: 16px;
+}
+
+
