+++ src/components/tag/textarea/textarea.css
@@ -1,0 +1,15 @@
+@import url('../../../styles/style.css');
+
+.form-item textarea {
+  padding: 8px;
+  margin: 6px 0;
+
+  border: 1px solid var(--grey-200);
+  border-radius: 8px;
+
+  font-size: 16px;
+}
+
+.form-item textarea {
+  resize: none;
+}
