+++ src/index.js
@@ -1,8 +1,8 @@
 import './styles/style.css';
 import './app.ts';
-import './components/Header.ts';
+import './components/header/Header';
 import './components/FilterContainer.ts';
-import './components/Modal.ts';
+import './components/modal/Modal';
 import './components/RestaurantForm.ts';
 
 import root from './Root';
